srtarted on login, refactored some stuff
