+++ front-end/src/api/api_wrapper.js
@@ -25,4 +25,9 @@ export default {
       return error.response;
     });
   },
+
+  authenticate(authToken) {
+    //for now, say everything is authenticated
+    return true;
+  },
 }
