+++ front-end/src/components/PageAddons/Footer.vue
@@ -19,7 +19,7 @@
     width: 105%;
     margin-top: 6%;
     /*background: linear-gradient(to top, #960d04 0%, #d10d00 100%);*/
-    background-image: url("../assets/background4.jpg");
+    background-image: url("../../assets/background4.jpg");
     background-repeat:no-repeat;
     background-size:cover;
   }
