+++ front-end/src/components/PageAddons/Header.vue
@@ -59,7 +59,7 @@
   header {
     position: relative;
     /*background: linear-gradient(to bottom, #960d04 0%, #d10d00 100%);*/
-    background-image: url("../assets/background4.jpg");
+    background-image: url("../../assets/background4.jpg");
     background-size: cover;
     margin: 0 0 5% 0;
     height: 10%;
