+++ front-end/src/components/UserView/CompactRestaurant.vue
