+++ front-end/src/components/UserView/DetailedRestaurant.vue
